+++ Mastermind/Ball.cs
@@ -1,0 +1,12 @@
+namespace Mastermind
+{
+      public class Ball
+    {
+        public string Letter {get; private set;}
+
+        public Ball (string letter)
+        {
+            this.Letter = letter;
+        }
+    }
+}
