+++ Mastermind/Mastermind.cs
@@ -85,42 +85,9 @@ namespace Mastermind
         }
     }
 
-    class Ball
-    {
-        public string Letter {get; private set;}
+  
 
-        public Ball (string letter)
-        {
-            this.Letter = letter;
-        }
-    }
+    
 
-    class Row
-    {
-        public Ball[] balls = new Ball[4];
-        public Row (Ball[] balls)
-        {
-            this.balls = balls;
-        } 
-
-        public string Balls{
-            get {
-                foreach (var ball in this.balls)
-                 {
-                     Console.WriteLine(ball.Letter);
-                }
-                return "";
-            }
-        }
-    }
-
-    class Game
-    {
-        private List<Row> rows = new List<Row> ();
-        private string[] answer = new string[4];
-        public Game(string[] answer)
-        {
-            this.answer = answer;
-        }
-    }
+    
 }
